+++ lib/tasks/decidim_app.rake
@@ -2,6 +2,7 @@
 
 require "decidim/admin_creator"
 require "decidim/system_admin_creator"
+require "k8s/configuration_exporter"
 
 namespace :decidim_app do
   desc "Setup Decidim-app"
@@ -36,12 +37,27 @@ namespace :decidim_app do
     Decidim::SystemAdminCreator.create!(ENV) ? puts("System admin created successfully") : puts("System admin creation failed")
   end
 
-  # This task is used to upgrade your decidim-app to the latest version
-  # Meant to be used in a CI/CD pipeline or a k8s job/operator
-  # You can add your own customizations here
-  desc "Upgrade decidim-app"
-  task upgrade: :environment do
-    puts "Running db:migrate"
-    Rake::Task["db:migrate"].invoke
+  namespace :k8s do
+    # This task is used to upgrade your decidim-app to the latest version
+    # Meant to be used in a CI/CD pipeline or a k8s job/operator
+    # You can add your own customizations here
+    desc "Upgrade decidim-app"
+    task upgrade: :environment do
+      puts "Running db:migrate"
+      Rake::Task["db:migrate"].invoke
+    end
+
+    desc "usage: bundle exec rails k8s:dump_db"
+    task dump_db: :environment do
+      K8s::OrganizationExporter.dump_db
+    end
+
+    desc "usage: bundle exec rails k8s:export_configuration IMAGE=<docker_image_ref>"
+    task export_configuration: :environment do
+      image = ENV["IMAGE"]
+      raise "You must specify a docker image, usage: bundle exec rails k8s:export_configuration IMAGE=<image_ref>" if image.blank?
+
+      K8s::ConfigurationExporter.export!(image)
+    end
   end
 end
